Test creation of JSON file from magnet info
Move the process of creating the JSON file to its own function to enable
testing of it. Mock the return value of the open() function with a
TemporaryFile. Change the file handling in id_setup.py to use a with
statement to make the mocking of the open() function easier, and also to
make the file handling safer.