Use test JSON data that generates a smaller .h5 file
The test_cpmu.json file is generated from the id_setup.py script, and
it's used by the lookup_generator.py script to generate a .h5 file.
Depending on what parameters are passed to id_setup.py, the JSON data
will then have a knock-on effect on what is in the .h5 file produced
from it, and in particular, how large the .h5 file produced is. Use a
test_cpmu.json file that will produce a .h5 file that is reasonably
sized to be put in the repo in the future.